Update tree code
