+++ nvflare/app_opt/xgboost/tree_based/shareable_generator.py
@@ -20,44 +20,7 @@ from nvflare.apis.shareable import Shareable
 from nvflare.app_common.abstract.model import ModelLearnable, ModelLearnableKey, model_learnable_to_dxo
 from nvflare.app_common.abstract.shareable_generator import ShareableGenerator
 from nvflare.app_common.app_constant import AppConstants
-
-
-def _get_xgboost_model_attr(xgb_model):
-    num_parallel_tree = int(
-        xgb_model["learner"]["gradient_booster"]["model"]["gbtree_model_param"]["num_parallel_tree"]
-    )
-    num_trees = int(xgb_model["learner"]["gradient_booster"]["model"]["gbtree_model_param"]["num_trees"])
-    return num_parallel_tree, num_trees
-
-
-def update_model(prev_model, model_update):
-    if not prev_model:
-        return model_update
-    else:
-        # Append all trees
-        # get the parameters
-        pre_num_parallel_tree, pre_num_trees = _get_xgboost_model_attr(prev_model)
-        cur_num_parallel_tree, add_num_trees = _get_xgboost_model_attr(model_update)
-
-        # check num_parallel_tree, should be consistent
-        if cur_num_parallel_tree != pre_num_parallel_tree:
-            raise ValueError(
-                f"add_num_parallel_tree should not change, previous {pre_num_parallel_tree}, current {add_num_parallel_tree}"
-            )
-        prev_model["learner"]["gradient_booster"]["model"]["gbtree_model_param"]["num_trees"] = str(
-            pre_num_trees + cur_num_parallel_tree
-        )
-        # append the new trees
-        append_info = model_update["learner"]["gradient_booster"]["model"]["trees"]
-        for tree_ct in range(cur_num_parallel_tree):
-            append_info[tree_ct]["id"] = pre_num_trees + tree_ct
-            prev_model["learner"]["gradient_booster"]["model"]["trees"].append(append_info[tree_ct])
-            prev_model["learner"]["gradient_booster"]["model"]["tree_info"].append(0)
-        # append iteration_indptr
-        prev_model["learner"]["gradient_booster"]["model"]["iteration_indptr"].append(
-            pre_num_trees + cur_num_parallel_tree
-        )
-        return prev_model
+from nvflare.app_opt.xgboost.tree_based.utils import update_model
 
 
 class XGBModelShareableGenerator(ShareableGenerator):
